refactor: Modal 컴포넌트 Content와 합친 후 변경사항 대응 (#490)
